refactor : VerticalTeamStatCards detph 낮추기 위해 분리및 결합했습니다
