+++ src/1_Page/Championship/ui/DashBoard/ui/MatchListTab/ui/MatchLineupContainer/ui/VerticalPersonStatCards/ui/TeamTable/ui/PlayerHistoryRow/type.d.ts
@@ -1,5 +1,5 @@
 type PlayerHistoryRowProps = {
-  p: PlayerStats;
+  player: PlayerStats;
   maxGoal: number;
   maxAssist: number;
   personEvidenceImage?: PlayerEvidenceImg[];
