+++ src/1_Page/Championship/ui/DashBoard/ui/MatchListTab/ui/MatchLineupContainer/ui/VerticalTeamStatCards/lib/getMatchDetailStats.ts
@@ -1,0 +1,17 @@
+export function getCurrentMomPlayer(
+  players: PlayerStats[] | undefined,
+  momIdx: number | undefined
+) {
+  return players?.find((p) => p.player_list_idx === momIdx);
+}
+
+export function getEvidenceUrls(
+  responseUrl: string[] | null,
+  evidenceImage: TeamEvidenceImg[] | undefined
+) {
+  return (
+    responseUrl ??
+    evidenceImage?.map((i) => i.match_team_stats_evidence_img) ??
+    []
+  );
+}
