--- src/1_Page/Championship/ui/DashBoard/ui/MatchListTab/ui/MatchLineupContainer/ui/VerticalTeamStatCards/ui/TeamStatCard/type.d.ts
@@ -1,3 +1,0 @@
-type TeamStatCardProps = {
-  teamData: TeamData;
-};
